HSEARCH-2919 Clearly separate JSR-352 entity manager factory retrieval ITs
They are very specific, so it should be clearer that way.
